requirements.txt
chardet==3.0.4
html5lib==1.1
idna==2.10
Pillow==7.2.0
PyPDF2==1.26.0
razorpay==1.2.0
reportlab==3.5.51
requests==2.24.0
stripe==2.50.0
urllib3==1.25.10
webencodings==0.5.1
xhtml2pdf==0.2.4
appdirs==1.4.4
asgiref==3.3.1
autopep8==1.5.5
certifi==2020.12.5
coverage==5.5
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
Faker==6.5.2
filelock==3.0.12
gunicorn==20.0.4
pipenv==2020.11.15
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
validate-email==1.3
whitenoise==5.2.0

